recruitment 검색 (이름, 주소)

+++ src/main/java/com/example/bob_friend/service/RecruitmentService.java
@@ -1,8 +1,10 @@
 package com.example.bob_friend.service;
 
+import com.example.bob_friend.model.dto.MemberDto;
 import com.example.bob_friend.model.dto.RecruitmentDto;
 import com.example.bob_friend.model.entity.Member;
 import com.example.bob_friend.model.entity.Recruitment;
+import com.example.bob_friend.model.exception.MemberNotAllowedException;
 import com.example.bob_friend.model.exception.RecruitmentIsFullException;
 import com.example.bob_friend.model.exception.RecruitmentNotActiveException;
 import com.example.bob_friend.model.exception.RecruitmentNotFoundException;
@@ -28,11 +30,13 @@ public class RecruitmentService {
         return new RecruitmentDto.Response(recruitment);
     }
 
+
     public List<RecruitmentDto.Response> findAll() {
         return recruitmentRepository.findAll().stream()
                 .map(RecruitmentDto.Response::new)
                 .collect(Collectors.toList());
     }
+
 
     public RecruitmentDto.Response createRecruitment(RecruitmentDto.Request recruitmentRequestDto) {
         Member currentMember = memberService.getCurrentMember();
@@ -42,9 +46,19 @@ public class RecruitmentService {
         return new RecruitmentDto.Response(savedRecruitment);
     }
 
+
     public void delete(Long recruitmentId) {
-        recruitmentRepository.deleteById(recruitmentId);
+        Member currentMember = memberService.getCurrentMember();
+        Recruitment recruitment = recruitmentRepository.findById(recruitmentId)
+                .orElseThrow(() -> {
+                    throw new RecruitmentNotFoundException(recruitmentId);
+                });
+        if (recruitment.getAuthor().equals(currentMember))
+            recruitmentRepository.deleteById(recruitmentId);
+        else
+            throw new MemberNotAllowedException(currentMember.getEmail());
     }
+
 
     public RecruitmentDto.Response update(Long recruitmentId,
                                           RecruitmentDto.Request update) {
@@ -58,13 +72,22 @@ public class RecruitmentService {
         return new RecruitmentDto.Response(savedRecruitment);
     }
 
-    public List<RecruitmentDto.Response> findAllByRestaurantNameOrRestaurantAddress(
+
+    public List<RecruitmentDto.Response> findAllByRestaurantNameAndRestaurantAddress(
             String restaurantName,
             String restaurantAddress) {
-        return recruitmentRepository.findAllByRestaurantNameOrRestaurantAddress(restaurantName,restaurantAddress).stream()
+        return recruitmentRepository.findAllByRestaurantNameAndRestaurantAddress(restaurantName, restaurantAddress).stream()
                 .map(RecruitmentDto.Response::new)
                 .collect(Collectors.toList());
     }
+
+
+    public List<RecruitmentDto.Response> findAllByRestaurantAddress(String restaurantAddress) {
+        return recruitmentRepository.findAllByRestaurantAddress(restaurantAddress).stream()
+                .map(RecruitmentDto.Response::new)
+                .collect(Collectors.toList());
+    }
+
 
     public List<RecruitmentDto.Response> findAllAvailableRecruitments() {
         Member currentMember = memberService.getCurrentMember();
@@ -76,6 +99,7 @@ public class RecruitmentService {
                 .collect(Collectors.toList());
     }
 
+
     public List<RecruitmentDto.Response> findAllJoinedRecruitments() {
         Member author = memberService.getCurrentMember();
         return recruitmentRepository.findAll().stream()
@@ -86,12 +110,14 @@ public class RecruitmentService {
                 .collect(Collectors.toList());
     }
 
+
     public List<RecruitmentDto.Response> findMyRecruitments() {
         Member author = memberService.getCurrentMember();
         return recruitmentRepository.findAllByAuthor(author).stream()
                 .map(RecruitmentDto.Response::new)
                 .collect(Collectors.toList());
     }
+
 
     @Transactional
     public RecruitmentDto.Response joinOrUnjoin(Long recruitmentId) throws RecruitmentIsFullException, RecruitmentNotActiveException {
@@ -111,6 +137,7 @@ public class RecruitmentService {
 
         return new RecruitmentDto.Response(recruitment);
     }
+
 
     private void validateRecruitment(Recruitment recruitment) {
         if (recruitment.isFull())
